Added anchor functionality to tabs
The page will now load with a given tab if it recognizes the anchor in the
url. Navigating tabs will now also set the appropraite anchor in the url.
Page navigation script was also cleaned/tweaked up a bit, I've tested most cases
with this new implementation, but edge cases may now behave differently,
for better or worse